fix problem with finding users city
before this, it would crash if the city wasn't set
yet.

+++ partEZ/database/migrations/2016_03_02_004702_add_city_to_users_table.php
@@ -1,6 +1,5 @@
 <?php
 
-use Illuminate\Database\Schema\Blueprint;
 use Illuminate\Database\Migrations\Migration;
 
 class AddCityToUsersTable extends Migration
@@ -27,9 +26,10 @@ class AddCityToUsersTable extends Migration
      */
     public function down()
     {
-        Schema::table('users', function($table)
-        {
-            $table->dropColumn('city');
-        });
+        if (Schema::hasColumn('users', 'city')) {
+            Schema::table('users', function ($table) {
+                $table->dropColumn('city');
+            });
+        }
     }
 }
